# smartgas-payments-composer/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Smart Gas API Specification",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/gas-stations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get paginated gas stations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gas Stations"
                ],
                "summary": "Gas Station List",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in gas stations, it looks up through id, name ip",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gas station paginated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GasStationListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create Gas Station",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gas Stations"
                ],
                "summary": "Gas Station Create",
                "parameters": [
                    {
                        "description": "Gas Station Body",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GasStationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Gas station create response",
                        "schema": {
                            "$ref": "#/definitions/dto.GasStationCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/gas-stations/all": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all gas stations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gas Stations"
                ],
                "summary": "Gas Station List All",
                "responses": {
                    "200": {
                        "description": "Gas stations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GasStationListAllResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/gas-stations/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Gas station detail",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gas Stations"
                ],
                "summary": "Gas Station Detail",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "uuid4 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gas station detail",
                        "schema": {
                            "$ref": "#/definitions/dto.GasStationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Gas station",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gas Stations"
                ],
                "summary": "Gas Station Update",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "uuid4 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Gas station values",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GasStationUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Updated",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "409": {
                        "description": "Duplicated entry",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BadRequestMessage": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GasStationCreateRequest": {
            "type": "object",
            "required": [
                "cre_permission",
                "external_id",
                "ip",
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "cre_permission": {
                    "type": "string",
                    "example": "PL/01/01..."
                },
                "external_id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.100.100"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "Guerrero"
                }
            }
        },
        "dto.GasStationCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "23ae8c18-4d7a-41a3-a148-8ae2d0a75690"
                }
            }
        },
        "dto.GasStationGetResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "cre_permission": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string",
                    "example": "13"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.100.100"
                },
                "name": {
                    "type": "string",
                    "example": "Guerrero"
                }
            }
        },
        "dto.GasStationListAllResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "23ae8c18-4d7a-41a3-a148-8ae2d0a75690"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.100.100"
                },
                "name": {
                    "type": "string",
                    "example": "Guerrero"
                }
            }
        },
        "dto.GasStationListResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "cre_permission": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string",
                    "example": "13"
                },
                "id": {
                    "type": "string",
                    "example": "23ae8c18-4d7a-41a3-a148-8ae2d0a75690"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.100.100"
                },
                "name": {
                    "type": "string",
                    "example": "Guerrero"
                }
            }
        },
        "dto.GasStationUpdateRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "cre_permission": {
                    "type": "string",
                    "example": "PL/01/01..."
                },
                "external_id": {
                    "type": "string",
                    "example": "12"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.100.100"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "Guerrero"
                }
            }
        },
        "dto.GeneralMessage": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "format": "array"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and paste the access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}